@@ -134,43 +132,43 @@
         "enzyme": "^3.10.0",
         "enzyme-adapter-react-16": "^1.10.0",
         "enzyme-to-json": "^3.3.4",
-        "eslint": "^7.7.0",
-        "eslint-import-resolver-typescript": "^2.2.1",
-        "eslint-import-resolver-webpack": "^0.12.2",
-        "eslint-plugin-import": "^2.22.0",
+        "eslint": "^8.34.0",
+        "eslint-import-resolver-typescript": "^3.5.3",
+        "eslint-import-resolver-webpack": "^0.13.2",
+        "eslint-plugin-import": "^2.27.5",
         "file-loader": "^1.1.11",
         "fork-ts-checker-webpack-plugin": "^6.0.0",
         "header-case-normalizer": "1.0.3",
         "html-webpack-plugin": "^4.3.0",
         "husky": "^1.0.0-rc.13",
-        "isomorphic-ws": "^4.0.1",
+        "isomorphic-ws": "^5.0.0",
         "javascript-obfuscator": "^2.6.0",
         "jest": "^26.4.1",
         "jest-config": "^26.4.1",
         "jest-environment-enzyme": "^7.1.2",
         "jest-enzyme": "^7.1.2",
-        "js-combinatorics": "0.5.0",
+        "js-combinatorics": "2.1.1",
         "lodash-webpack-plugin": "^0.11.6",
-        "mini-css-extract-plugin": "^0.10.0",
+        "mini-css-extract-plugin": "^1.0.0",
         "npm-run-all": "^4.1.5",
         "optimize-css-assets-webpack-plugin": "^5.0.3",
-        "postcss": "^8.1.4",
-        "postcss-autoreset": "^2.0.2",
-        "postcss-custom-properties": "^10.0.0",
-        "postcss-hexrgba": "^2.0.1",
-        "postcss-import": "^13.0.0",
+        "postcss": "^8.4.21",
+        "postcss-autoreset": "^3.0.4",
+        "postcss-custom-properties": "^13.1.3",
+        "postcss-hexrgba": "^2.1.0",
+        "postcss-import": "^15.0.0",
         "postcss-loader": "^4.0.4",
-        "postcss-nested": "^5.0.1",
-        "ramda": "^0.26.1",
-        "react-hot-loader": "^4.3.4",
-        "react-test-renderer": "^16.4.2",
-        "redux-mock-store": "^1.5.3",
-        "rimraf": "^2.6.2",
+        "postcss-nested": "^6.0.0",
+        "ramda": "^0.28.0",
+        "react-hot-loader": "^4.13.1",
+        "react-test-renderer": "^18.2.0",
+        "redux-mock-store": "^1.5.4",
+        "rimraf": "^4.1.2",
         "sass-loader": "10.2.0",
-        "source-map-explorer": "^1.7.0",
+        "source-map-explorer": "^2.5.3",
         "speed-measure-webpack-plugin": "^1.3.3",
-        "stylelint-config-prettier": "^8.0.2",
-        "stylelint-config-recommended": "^3.0.0",
+        "stylelint-config-prettier": "^9.0.5",
+        "stylelint-config-recommended": "^10.0.0",
         "tailwindcss": "^2.2.7",
         "ts-jest": "^29.0.5",
         "ts-node": "^10.9.1",
